Move runChangesConcurrent to defaultMethod
This commit moves the runChangesConcurrent definition
to the DefaultMethod struct so every method does not
have to define it.
