+++ pkg/engine/kube.go
@@ -99,25 +99,6 @@ func (k *Kube) Apply(ctx, conn context.Context, target *Target, currentState, de
 	return nil
 }
 
-func (k *Kube) runChangesConcurrent(ctx context.Context, conn context.Context, changeMap map[*object.Change]string) error {
-	ch := make(chan error)
-	for change, changePath := range changeMap {
-		go func(ch chan<- error, changePath string, change *object.Change) {
-			if err := k.MethodEngine(ctx, conn, change, changePath); err != nil {
-				ch <- utils.WrapErr(err, "error running engine method for change from: %s to %s", change.From.Name, change.To.Name)
-			}
-			ch <- nil
-		}(ch, changePath, change)
-	}
-	for range changeMap {
-		err := <-ch
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 func (k *Kube) kubePodman(ctx, conn context.Context, path string, prev *string) error {
 	if path != deleteFile {
 		klog.Infof("Creating podman container from %s using kube method", path)
